requirements.txt
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
APScheduler==3.11.0
async-timeout==5.0.1
attrs==25.3.0
bcrypt==4.3.0
beautifulsoup4==4.13.3
bs4==0.0.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cssselect==1.3.0
exceptiongroup==1.2.2
fake-useragent==2.1.0
fastapi==0.115.11
filelock==3.18.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
joblib==1.4.2
lxml==5.3.1
lxml_html_clean==0.4.1
markovify==0.9.4
nest-asyncio==1.6.0
numpy==2.2.4
orjson==3.10.15
outcome==1.3.0.post0
parse==1.20.2
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyee==11.1.1
PyJWT==2.10.1
pyppeteer==2.0.0
pyquery==2.0.1
PySocks==1.7.1
python-telegram-bot==22.0
redis==5.2.1
requests==2.32.3
requests-html==0.10.0
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.30.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
starlette==0.46.1
threadpoolctl==3.6.0
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
Unidecode==1.3.8
urllib3==1.26.20
uvicorn==0.34.0
w3lib==2.3.1
websocket-client==1.8.0
websockets==10.4
wsproto==1.2.0
zipp==3.21.0
